Move HdfsUrl Spark reader and writer ops to core package object
Permits to opt out of having them in the scope when extending SparkJob
and allows to convert them to value classes as suggested in #1.
